requirements.txt
requests==2.31.0
pyyaml==6.0.1
pytest==7.4.3
click==8.1.7
flask==3.0.0
